refactor(fluentui-publish): de-chunk nx-publish into proper src/* module scaffold
