+++ scripts/fluentui-publish/bin/northstar-release.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node
+
+// @ts-check
+
+const { joinPathFragments } = require('@nx/devkit');
+const { registerTsProject } = require('@nx/js/src/internal');
+
+registerTsProject(joinPathFragments(__dirname, '..', 'tsconfig.lib.json'));
+
+const { main } = require('../src/cli');
+
+main();
