+++ scripts/fluentui-publish/package.json
@@ -2,7 +2,9 @@
   "name": "@fluentui/scripts-fluentui-publish",
   "version": "0.0.1",
   "private": true,
-  "main": "cli.js",
+  "bin": {
+    "northstar-release": "./bin/northstar-release.js"
+  },
   "scripts": {
     "format": "prettier -w --ignore-path ../../.prettierignore .",
     "format:check": "yarn format -c",
