+++ scripts/fluentui-publish/project.json
@@ -1,7 +1,7 @@
 {
   "name": "@fluentui/scripts-fluentui-publish",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "scripts/fluentui-publish",
+  "sourceRoot": "scripts/fluentui-publish/src",
   "projectType": "library",
   "tags": ["tools"]
 }
