+++ scripts/fluentui-publish/tsconfig.lib.json
@@ -7,5 +7,5 @@
     "types": ["node"]
   },
   "exclude": ["**/*.spec.ts", "**/*.test.ts"],
-  "include": ["./**/*.ts", "./**/*.js"]
+  "include": ["./src/**/*.ts", "./src/**/*.js"]
 }
